// src/modules/fit-analysis/fit-analysis.types.ts
export interface TopMeasurements {
  bustCm: number;
  shoulderCm?: number;
  sleeveCm?: number;
  lengthCm: number;
}

export interface TrouserMeasurements {
  waistCm: number;
  hipsCm: number;
  inseamCm: number;
  riseCm: number;
}

export interface DressMeasurements {
  bustCm: number;
  waistCm: number;
  hipsCm: number;
  lengthCm: number;
}